combined interfaces into /types/index.d.ts
