--- heat-stack/model/Case.tsx
@@ -1,4 +1,0 @@
-export interface Case {
-	firstName: string
-	lastName: string
-}
