--- heat-stack/model/Home.tsx
@@ -1,12 +1,0 @@
-export interface Home {
-	livingArea: number
-	fuelType: 'Natural Gas' | 'Oil' | 'Propane'
-	designTemperatureOverride: number
-	heatingSystemEfficiency: number
-	thermostatSetPoint: number
-	setbackTemperature: number
-	setbackHoursPerDay: number
-	numberOfOccupants: number
-	estimatedWaterHeatingEfficiency: number
-	standByLosses: number
-}
