--- heat-stack/model/Location.tsx
@@ -1,8 +1,0 @@
-export interface Location {
-	address: string
-	addressLine2: string
-	city: string
-	state: string
-	zip: string
-	country: string
-}
