--- heat-stack/model/NaturalGasBill.tsx
@@ -1,3 +1,0 @@
-export interface NaturalGasBill {
-	provider: string
-}
