--- heat-stack/model/NaturalGasBillRecord.tsx
@@ -1,6 +1,0 @@
-export interface NaturalGasBillRecord {
-	periodStartDate: Date
-	periodEndDate: Date
-	usageTherms: number
-	inclusionOverride: 'Include' | 'Do not include' | 'Include in other analysis'
-}
