--- heat-stack/model/OilPropaneBill.tsx
@@ -1,4 +1,0 @@
-export interface OilPropaneBill {
-	provider: string
-	precedingDeliveryDate: Date
-}
