--- heat-stack/model/OilPropaneBillRecord.tsx
@@ -1,6 +1,0 @@
-export interface OilPropaneBillRecord {
-	periodStartDate: Date
-	periodEndDate: Date
-	gallons: number
-	inclusionOverride: 'Include' | 'Do not include' | 'Include in other analysis'
-}
